# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PlayOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /host/login:
    post:
      operationId: hostLogin
      tags: [hosts]
      summary: Login as a host (phone + password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostLoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /host/apply:
    post:
      operationId: applyAsHost
      tags: [hosts]
      summary: Submit a host application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostApplicationBody"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==================== GAMES ====================
  /games:
    get:
      operationId: listGames
      tags: [games]
      summary: List all public open games
      parameters:
        - name: city
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameSummary"

    post:
      operationId: createGame
      tags: [games]
      summary: Create a new game (organiser only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGameBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /games/{id}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get game details including bookings/players
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateGame
      tags: [games]
      summary: Update game (organiser only, own game)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGameBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteGame
      tags: [games]
      summary: Delete game (organiser only, own game)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==================== BOOKINGS ====================
...
  /games/{id}/manage:
    get:
      operationId: getGameManagement
      tags: [games]
      summary: Get full management view of a game (organiser only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameManagement"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pitches:
    get:
      operationId: listPitches
      tags: [hosts]
      summary: Get organiser's pitches
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pitch"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createPitch
      tags: [hosts]
      summary: Create a new pitch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePitchBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pitch"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ==================== AUTH SCHEMAS ====================
...
    HostLoginBody:
      type: object
      properties:
        phone:
          type: string
        password:
          type: string
      required:
        - phone
        - password
    HostApplicationBody:
      type: object
      properties:
        name:
          type: string
        pitchName:
          type: string
        phone:
          type: string
        city:
          type: string
      required:
        - name
        - pitchName
        - phone
        - city
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        name:
          type: string
        role:
          type: string
          enum: [player, organiser, host, admin]
        token:
          type: string
          description: JWT auth token returned on login/signup only
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - role
        - createdAt

    # ==================== GAME SCHEMAS ====================
    Game:
      type: object
      properties:
        id:
          type: string
        organiserId:
          type: string
        title:
          type: string
        pitchName:
          type: string
        locationText:
          type: ["string", "null"]
        kickoffTime:
          type: string
          format: date-time
        price:
          type: number
        capacity:
          type: integer
        status:
          type: string
          enum: [open, full, cancelled]
        autoCancelHours:
          type: integer
        durationMinutes:
          type: integer
        isPublic:
          type: boolean
        mapsUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - organiserId
        - title
        - pitchName
        - kickoffTime
        - price
        - capacity
        - status
        - autoCancelHours
        - durationMinutes
        - isPublic
        - createdAt
    GameSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        pitchName:
          type: string
        locationText:
          type: ["string", "null"]
        kickoffTime:
          type: string
          format: date-time
        price:
          type: number
        capacity:
          type: integer
        status:
          type: string
          enum: [open, full, cancelled]
        bookedCount:
          type: integer
        durationMinutes:
          type: integer
        isPublic:
          type: boolean
        mapsUrl:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
      required:
        - id
        - title
        - pitchName
        - kickoffTime
        - price
        - capacity
        - status
        - bookedCount
        - durationMinutes
        - isPublic
    GameDetail:
      type: object
      properties:
        id:
          type: string
        organiserId:
          type: string
        title:
          type: string
        pitchName:
          type: string
        locationText:
          type: ["string", "null"]
        kickoffTime:
          type: string
          format: date-time
        price:
          type: number
        capacity:
          type: integer
        status:
          type: string
          enum: [open, full, cancelled]
        autoCancelHours:
          type: integer
        durationMinutes:
          type: integer
        isPublic:
          type: boolean
        mapsUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/BookingWithPlayer"
        bookedCount:
          type: integer
      required:
        - id
        - organiserId
        - title
        - pitchName
        - kickoffTime
        - price
        - capacity
        - status
        - autoCancelHours
        - durationMinutes
        - isPublic
        - createdAt
        - bookings
        - bookedCount
    GameManagement:
      type: object
      properties:
        game:
          $ref: "#/components/schemas/Game"
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/BookingWithPlayer"
        bookedCount:
          type: integer
        netPayout:
          type: number
        shareUrl:
          type: string
      required:
        - game
        - bookings
        - bookedCount
        - netPayout
        - shareUrl
    CreateGameBody:
      type: object
      properties:
        title:
          type: string
        pitchName:
          type: string
        locationText:
          type: ["string", "null"]
        kickoffTime:
          type: string
        price:
          type: number
        capacity:
          type: integer
        autoCancelHours:
          type: integer
        durationMinutes:
          type: integer
        isPublic:
          type: boolean
        mapsUrl:
          type: ["string", "null"]
      required:
        - title
        - pitchName
        - kickoffTime
        - price
        - capacity
    UpdateGameBody:
      type: object
      properties:
        title:
          type: string
        pitchName:
          type: string
        locationText:
          type: ["string", "null"]
        kickoffTime:
          type: string
        price:
          type: number
        capacity:
          type: integer
        isPublic:
          type: boolean
        status:
          type: string
          enum: [open, full, cancelled]
        mapsUrl:
          type: ["string", "null"]

    # ==================== MY BOOKINGS SCHEMAS ====================
...
    BookingWithPlayer:
      type: object
      properties:
        id:
          type: string
        gameId:
          type: string
        userId:
          type: string
        team:
          type: integer
        slotIndex:
          type: integer
        paymentStatus:
          type: string
          enum: [pending, paid, refunded]
        paymentId:
          type: ["string", "null"]
        bookedAt:
          type: string
          format: date-time
        playerName:
          type: string
        playerPhone:
          type: ["string", "null"]
      required:
        - id
        - gameId
        - userId
        - team
        - slotIndex
        - paymentStatus
        - bookedAt
        - playerName
...
    Pitch:
      type: object
      properties:
        id:
          type: string
        organiserId:
          type: string
        name:
          type: string
        mapsUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - organiserId
        - name
        - createdAt
    CreatePitchBody:
      type: object
      properties:
        name:
          type: string
        mapsUrl:
          type: ["string", "null"]
      required:
        - name